pyproject.toml
[tool.poetry]
name = "django-rest"
version = "0.1.0"
description = "Build a REST API in 30 minutes with Django REST Framework."
authors = ["Marco André Mendes <marcoandre@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
django = "^3.0.6"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.2"
isort = "^4.3.21"
pre-commit = "^2.4.0"

[tool.black]
line-length = 119

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

